+++ Filter/read_length_calc.cpp
@@ -8,7 +8,7 @@
 #include <api/BamReader.h>
 
 
-std::vector <std::string> fastq_count ( std::string infile , int max_count){
+std::vector <std::string> fastq_count ( std::string infile , int64_t max_count){
   std::vector < std::string > output; 
   std::ifstream fin (infile.c_str());
   std::string id;
@@ -16,7 +16,7 @@ std::vector <std::string> fastq_count ( std::string infile , int max_count){
   std::string desc;
   std::string qual;
 
-  int count=0 ;
+  int64_t count=0 ;
 
   while ( getline ( fin , id ) && getline ( fin , seq ) && getline ( fin , desc ) && getline ( fin , qual ) && count < max_count){
     output.push_back ( seq );
@@ -27,7 +27,7 @@ std::vector <std::string> fastq_count ( std::string infile , int max_count){
   return output;
 }
 
-std::vector <std::string> fastq_gz_count ( std::string infile , int max_count){
+std::vector <std::string> fastq_gz_count ( std::string infile , int64_t max_count){
   std::vector < std::string > output;
   igzstream fin;
   fin.open(infile.c_str());
@@ -36,7 +36,7 @@ std::vector <std::string> fastq_gz_count ( std::string infile , int max_count){
   std::string desc;
   std::string qual;
 
-  int count=0 ;
+  int64_t count=0 ;
 
   while ( getline ( fin , id ) && getline ( fin , seq ) && getline ( fin , desc ) && getline ( fin , qual ) && count < max_count){
     output.push_back ( seq );
@@ -48,7 +48,7 @@ std::vector <std::string> fastq_gz_count ( std::string infile , int max_count){
 }
 
 
-std::vector <std::string> bam_count ( std::string infile , int max_count){
+std::vector <std::string> bam_count ( std::string infile , int64_t max_count){
   std::vector < std::string > output;
 
   BamTools::BamReader reader;
@@ -56,7 +56,7 @@ std::vector <std::string> bam_count ( std::string infile , int max_count){
 
   reader.Open ( infile );
   
-  int count = 0 ;
+  int64_t count = 0 ;
 
   while ( reader.GetNextAlignment(al) && count < max_count ){
     output.push_back(al.QueryBases);
@@ -75,32 +75,32 @@ int main ( int argc , char ** argv ){
   }
 
   std::string infile = argv[1];
-  int max_count = atoi ( argv[2] );
+  int64_t max_count = atoi ( argv[2] );
 
-  int read_length_sum =0 ;
-  int read_length_average =0 ;
-  int count = 0;
+  int64_t read_length_sum =0 ;
+  int64_t read_length_average =0 ;
+  int64_t count = 0;
 
   std::string id;
   std::string seq;
   std::string desc;
   std::string qual;
 
-  std::vector < std::string > (*fcnPtr)(std::string, int);
+  std::vector < std::string > (*fcnPtr)( std::string, int64_t );
   std::vector < std::string > seq_vec;
 
-  int Size = infile.size();
+  int64_t Size = infile.size();
 
-  int size9 = Size - 9;
+  int64_t size9 = Size - 9;
   if (size9 < 0 ) size9=0;
 
-  int size6 = Size - 6;
+  int64_t size6 = Size - 6;
   if (size6 < 0 ) size6=0;
 
-  int size4 = Size - 4;
+  int64_t size4 = Size - 4;
   if (size4 < 0 ) size4=0;
 
-  int size3 = Size - 3;
+  int64_t size3 = Size - 3;
   if (size3 < 0 ) size3=0;
 
   if ( infile.substr(size6) == ".fastq" ||
